Create LICENSE
updating the license to a more the more specific Apache2 License. The reason we want Apache2 is that apache2 makes explicit the contribution of patents and the github inbound=outbound means that one is affirming that license in the contribution. Whereas MIT gives wiggle room for someone to make a contribution and assert they didn't understand that it included a patent license. Apache2 makes all pull requests explicit contributions under the same terms. Whereas MIT makes no mention of how contributions by others are to be treated. So its less precise.